+++ lbz/consts.py
@@ -1,0 +1,1 @@
+DIRECT_LAMBDA_REQUEST = "direct_lambda_request"