+++ lbz/dev/test.py
@@ -83,7 +83,7 @@ class Client:
                     query_params[key] = [str(value)]
                 else:
                     query_params[key] = [str(elem) for elem in value]
-        return self.resource(
+        res: Response = self.resource(
             APIGatewayEvent(
                 resource_path=path,
                 method=method,
@@ -93,3 +93,4 @@ class Client:
                 headers=headers,
             )
         )()
+        return res